Fix placed head drops
When a head was placed, it would lose some metadata, so would lack the display name and could no longer be stacked with other matching heads.
This solution is a little overkill, in that it creates a new head and spawns that, but it also covers bases on any similar discrepancies that could occur.